Dictionary_Operations

Documentation

load_dictionary_without_frequencies :: String -> [String]

This function is used to parse the dictionary string ignoring it's frequency

load_dictionary_with_frequencies :: String -> [(String, Int)]

This function is used to parse the dictionary along with its frequency

give_me_words :: String -> [String]

It return the list of all words present in the string passed to it

give_me_word_frequency :: String -> [(String, Int)]

It gives the list of all distinct words in a string along with its frequency that is How many times that word occoured in the string

group_same_words :: String -> [[String]]

It gives the list of all words present in the string in which all identical words are grouped together into a nested list

clean_text :: [Char] -> [Char]

This function takes a string and replaces characters which are not allowed by char_filter with space