Main

Documentation

dispatch :: [(String, [String] -> IO ())]

This function enable to run the program in different ways when the arguments passed are different

main :: IO ()

t :: [String] -> IO ()

This is used to run the program when only the text file is specified by the user

td :: [String] -> IO ()

This is used to run the program when both dictionay and text file is specified by the user

tds :: [String] -> IO ()

This is used to run the program when dictionary , text file and suggestion number is specified by the user

te :: [String] -> IO ()

This is used to run the rogram only when text file and edit bound is specified by the user

tde :: [String] -> IO ()

This is used when the user specifies textfile, dictionary and the edit bound

tdse :: [String] -> IO ()

This is used to run the program when all the parameters are specified by the user

ts :: [String] -> IO ()

This is used to run the program when only text file and suggestion_no is specified by the user

tse :: [String] -> IO ()

This is used to run the program when textfile and edit bound and suggestin no is specified by the user