String_Operation

Documentation

wrong_words_of_string :: String -> String -> [String]

This Function takes a text line and dictionary as it's arguments and produces a list of all wrong words in that line.

indices_in_text :: (Num b, Enum b) => [Char] -> [Char] -> [b]

This function take a word and a text line as it's arguments and produces a list of position of that word in that line.

wrong_words_in_line :: String -> String -> [String]

This Function takes a text line and dictionary as it's arguments and produces a list of unique wrong words in that line.

position_of_wrong_words_in_line :: (Ord a, Num a, Enum a) => [Char] -> String -> [(a, [Char])]

This function takes the line and dictionary as it's argument and produces the list of pair of position of wrong word in that line and wrong word.

error_in_file :: (Ord t, Num t, Num t1, Enum t, Enum t1) => String -> String -> [(t1, t, [Char])]

This function takes a text string and dictionary string as it's argument and produces the list of tuples of line_number, position of worng word in line and the wrong word in the text.

make_tuple :: (Ord t, Num t, Enum t) => t1 -> [Char] -> String -> [(t1, t, [Char])]

This function takes line_number, line string and dictionary as it's arguments and produces a list of tuples of line_number, position of worng word in line and the wrong word in the text.

print_suggestion :: String -> String -> Int -> Int -> String

This is the main function that we call in Main.hs, It gives us the String of wrong words and suggestion of them in required format.

suggestions :: (Ord t, Num t, Num t1, Enum t, Enum t1) => String -> String -> Int -> Int -> [(t1, t, [Char], [[Char]])]

This function takes the text string, dictionary string , number of maximumm suggestion and edit distance as it's argument and produces the list of tuple of line number, position number, wrong word and suggestion for this wrong word for all wrong word preesent in given text string respective to the dictionary string.

give_me_suggestion :: [Char] -> String -> Int -> Int -> [[Char]]

This function take a word, a dictionary string, maximum number of suggestion and edit distance as it's argument and produces the list of suggestion for that wrong word according to the given dictionary.